requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
markdown2>=2.4.10
jsonschema>=4.21.1
astunparse>=1.6.3
tqdm>=4.66.2    
libmagic
chromadb>=0.4.24
ollama>=0.1.6
python-magic>=0.4.27